+++ CHANGELOG.rst
@@ -2,12 +2,14 @@
 Changelog for package grepros
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-1.1.0 (2023-12-27)
+1.1.0 (2023-12-28)
 -------------------
 * add support for splitting files in bag/HTML/MCAP/SQLite output
 * add api.to_duration()
 * do not use partial match for topic and typename filters without wildcards
 * support patterns yielding zero-length matches like "(?!not_this)"
+* support giving dashed names like "rollover-size" in format-specific write options
+  as underscored "rollover_size" instead
 * disallow unknown command-line flags and partial abbreviations
 * fix matching nested message values for patterns using start or end flags
 * fix ISO datetime support in earliest and latest timestamp arguments